third_party/WebKit/LayoutTests/external/wpt/service-workers/service-worker/fetch-request-resources.https-expected.txt

This is a testharness.js-based test.
FAIL Verify FetchEvent for resources. assert_equals: integrity of Script load (url:https://web-platform.test:8444/service-workers/service-worker/resources/dummy?test21) must be      . expected "     " but got ""
Harness: the test ran to completion.
